Add files via upload
I realize this is not compete. I have tried for 3 days and I can not solve the final two parts and I made the decision that it would be better to submit a partially complete assignment than to not complete/submit anything. Please provide any feedback so that I can learn from this experience for the next challenge